gsmi: Remove autoselected dependency on EFI and EFI_VARS
Instead of selecting EFI and EFI_VARS automatically when GSMI is
enabled let that portion of the driver be conditionally compiled
if EFI and EFI_VARS are enabled.

This allows the rest of the driver (specifically event log) to
be used if EFI_VARS is not enabled.

BUG=chromium-os:32188
TEST=manual: compile and run kernel on multiple devices

1) verify that EFI_VARS is not automatically selected when
CONFIG_GOOGLE_GSMI is enabled
2) verify that the kernel boots on Link and that GSMI event log
is still available and functional
3) specifically boot the kernel on Alex to ensure it does not
try to load efivars and that gsmi also does not load because it
is not in the supported DMI table

Change-Id: I6f9149b575976316dd1d4d8ac20f2688d2208767
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26273
Reviewed-by: Olof Johansson <olofj@chromium.org>

[bleung: fixup for 3.10 rebase]
Signed-off-by: Benson Leung <bleung@chromium.org>

[benzh: 3.14 rebase. In 3.14 GOOGLE_SMI depends on EFI, so remove EFI]
Signed-off-by: Ben Zhang <benzh@chromium.org>

Conflicts:
	drivers/firmware/google/gsmi.c
[filbranden: 3.18 rebase. Solved conflict related to the following commit:
 b12b73f1ca77 firmware: fix google/gsmi duplicate efivars_sysfs_init()]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>