CHROMIUM: Input: atmel_mxt_ts - add regulator control support
Allow the driver to optionally manage enabling/disable power to the touch
controller itself if regulators are defined in DTS. Note that for power
management we still use the old scheme and do not turn off the regulators
even if they are defined, we'll do that in a subsequent change.

The change is reworked from patch by Nick Dyer <nick.dyer@itdev.co.uk>

BUG=chrome-os-partner:35629
TEST=Boot and verify that touchscreen works; suspend and resume and
verify it continues to work.

Change-Id: I4a2f12f78608193432e3246d14d9ae5cc2adbb86
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243404
Reviewed-by: Douglas Anderson <dianders@chromium.org>

[dtor: 4.4 rebase - devm_gpiod_flags are now mandatory]