CHROMIUM: dark resume: Expose dark resume activation
Add a way for userspace to activate dark resume specifically for devices. This
will allow devices to specify to the kernel whether it should do something
different for a device (such as not come up at all) during a dark resume. This
allows drivers that run very different hardware to do something different based
on specific configurations for a machine.

BUG=chrome-os-partner:9812
TEST=suspend system, watch dark resumes happen

Change-Id: If2a6fab5500027d17347552b5a08b24d764aa946
[sonnyrao: 3.8: fixup conflict from added wakelock stuff]
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42923
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>

[filbranden: 3.18 rebase. Conflicts renamed {en,dis}abled -> _{en,dis}abled.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>