CHROMIUM: Event log support in the chromeos_arm platform driver.
An in memory buffer is set up covering a range of addresses specified in the
device tree, and when a panic happens, an event log event is generated and
stored in that buffer. When the system reboots as a result of the panic, the
firmware discovers the event and records it in the event log in flash.

I've also added a remove function which may or may not be desirable.

BUG=chrome-os-partner:16132
TEST=Built and booted on pit, triggered a panic using /proc/sysrq-trigger, and
after the reboot, used mosys eventlog to verify that a panic event had been
recorded.

Change-Id: I93d7c71d99c5a536aa78c59e8d9ade9a3525a0c5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/65154
Reviewed-by: Olof Johansson <olofj@chromium.org>