CHROMIUM: power: Clear dark_resume_state when using pm_test
Until selective resume is ready, the chrome os power manager will use
pm_test to indicate that it wants to transition from a dark resume to
fully resumed.  When this happens, we want to clear dark_resume_state so
that all devices will go through the proper resume path.

BUG=chromium:397777
TEST=manual
Tapped the keyboard while the system was in dark resume, checked the
power manager logs to make sure that the transition message appeared,
and verified that the usb devices came back up.

Change-Id: I4e79014b3098985b762961082a9b06c7b260376d
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/211519
Reviewed-by: Derek Basehore <dbasehore@chromium.org>

Conflicts:
	kernel/power/suspend.c
[rebase44(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>