CHROMIUM: sched: Energy-aware wake-up task placement
Let available compute capacity and estimated energy impact select
wake-up target cpu when energy-aware scheduling is enabled and the
system in not over-utilized (above the tipping point).

energy_aware_wake_cpu() attempts to find group of cpus with sufficient
compute capacity to accommodate the task and find a cpu with enough spare
capacity to handle the task within that group. Preference is given to
cpus with enough spare capacity at the current OPP. Finally, the energy
impact of the new target and the previous task cpu is compared to select
the wake-up target cpu.

BUG=chromium:580187
TEST=boot kevin/gru

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

Contains the following squashed commits:
"sched: Determine the current sched_group idle-state"
"sched: Estimate energy impact of scheduling decisions"
"sched: Extend sched_group_energy to test load-balancing decisions"
"sched: Calculate energy consumption of sched_group"

This is to fix warnings during compilation due to unused static
functions that are not used until this patch is applied.

Change-Id: Ie290971858d28c0c764d037366edd0386163476a
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/358231
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Javi Merino <javi.merino@arm.com>