CHROMIUM: usb_pd_charger: Set proper online status when port is a source
Set online status to 0 when the port is a source (ex. when a peripheral
is attached, or when the port is providing a charge). Though I can find
no spec regarding POWER_SUPPLY_PROP_ONLINE, this change seems consistent
with its usage in power_supply_core __power_supply_am_i_supplied() --
online status should be zero if the port isn't supplied with a charge.

BUG=chrome-os-partner:36650
TEST=Manual on Samus. Plug USB keyboard into port, verify that online
status is 0. Plug USB charger into port, verify that online status goes
to 1.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Change-Id: Icec8e8e1dcdaacffc2bf2463e729a2f31f43de7f
Reviewed-on: https://chromium-review.googlesource.com/248930
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>