CHROMIUM: atmel_mxt_ts - Disable hover on suspend
We don't want hover to wake the system when suspended,
so lets stash the value of the hover control register and
restore it on resume.

This somewhat straightforward change, unfortunately, runs
afoul of a limitation of the atmel firmware, specifically that
it forces a recalibration when certain registers are changed.

The recalibration itself is not a problem, but the two interrupts
that are fired (one at start and one at end) are. They all but
guarantee that the system will immediately wake from S3 on an idle
suspend.

This CL reuses the auto_cal_completion that was previously used
by mxt_suspend_enable_T9 to wait for the calbiration to finish.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:36837
TEST=powerd_dbus_suspend
Check that the system does not wake immediately.
Hover finger right above the pad. Check that the system does
not wake.
Touch the pad, check the system wakes

Change-Id: I210d66543755fe739d79faa2f07ed2cd172c6c9e
Reviewed-on: https://chromium-review.googlesource.com/250623
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Trybot-Ready: Benson Leung <bleung@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>