CHROMIUM: iwl7000: mvm: implement Tx latency threshold trigger
When mac80211 notices that a Tx packet's latency crossed a certain
threshold that was pre-configured, it notifies the driver.

The driver triggers an event to get the usniffer logs, and writes the
metadata to ftrace

There are 3 different modes for collecting the:
1. External buffer: used when you have an external device that save the
logs, the usniffer logs are not collected in this case, but a marker is
sent to the fw to mark the location of the event.
2. Immediate Internal buffer: used when the buffer on the device is small.
The driver triggers on the first packet that crosses the threshold, and
then waits one second before triggering again.
3. Delayed Internal buffer: used when the buffer on the device is big.
The driver triggers to collect the logs on the packet with the max
latency during the configured window

Add trigger for Tx latency packets that crossed the configured threshold.

Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
iwl7000-tree: ebe85149f7c21950d8766a7fce4c83fb909693b3