FROMLIST: device core: add device_is_bound()
Adds a function that tells whether a device is already bound to a
driver.

This is needed to warn when there is an attempt to change the PM domain
of a device that has finished probing already. The reason why we want to
enforce that is because in the general case that can cause problems and
also that we can simplify code quite a bit if we can always assume that.

BUG=chrome-os-partner:46491
TEST=boot glados

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
(am from https://patchwork.kernel.org/patch/7328221/)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I02b842a5ce3beca1d3228243bed64ad9b9a7d7d3
Reviewed-on: https://chromium-review.googlesource.com/305813
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>