CHROMIUM: iwl7000: mac80211: allow disabling U-APSD workarounds
U-APSD is disabled in mac80211 if the AP is SISO or non-HT, in order
to work around broken iPhones advertising U-APSD capability but not
really implementing it.

However, in certain situations it is necessary to disable this and
work with U-APSD on such APs; allow doing that through debugfs.

Signed-off-by: Avri Altman <avri.altman@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3770c5c0aa574fd863badbf5335b99cc4167a59a