CHROMIUM: Add generic way to find wakeup_type
This adds a generic way to get the wakeup_type for a system. It works if the
firmware doesn't get in the way of the kernel receiving interrupts.

This only works for wakeup_sources that are attached to a device.

BUG=chrome-os-partner:41007
TEST=on glados:
echo automatic > /sys/class/rtc/rtc0/device/power/wakeup_type
echo +10 > /sys/class/rtc/rtc0/wakealarm; powerd_dbus_suspend
Check that /sys/power/wakeup_type is automatic after the system wakes from
suspend

Change-Id: I2ccd2e49cf2bc37d36b342f613562200f299319c
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309001
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>