UPSTREAM: serial: support 16-bit register interface for console
Currently, 8-bit (MMIO) and 32-bit (MMIO32) register interfaces are
supported for the 8250 console, but the 16-bit (MMIO16) is not.
The 8250 UART device on my board is connected to a 16-bit bus and
my main motivation is to use earlycon with it.
(Refer to arch/arm/boot/dts/uniphier-support-card.dtsi)

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Reviewed-by: Peter Hurley <peter@hurleysoftware.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit bd94c4077a0b2ecc35562c294f80f3659ecd8499)

BUG=chromium:636585
TEST=see early prints on rk3399 with earlycon in kernel cmdline

Change-Id: Ic2cb71e778e0c019ba8aaa2c82bf76828b481445
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/367910
Reviewed-by: Guenter Roeck <groeck@chromium.org>