UPSTREAM: PM / devfreq: Set the freq_table of devfreq device
This patch initialize the freq_table array of each devfreq device by using
the devfreq_set_freq_table(). If freq_table is NULL, the devfreq framework
is not able to support the frequency transtion information through sysfs.

The OPP core uses the integer type for the number of opps in the opp list
and uses the 'unsigned long' type for each frequency. So, this patch
modifies the type of some variable as following:
- the type of freq_table : unsigned int -> unsigned long
- the type of max_state  : unsigned int -> int

- Corrected types, format strings, mutex usages by MyungJoo

BUG=chrome-os-partner:42645
TEST=Build and boot on smaug.

Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>
(cherry picked from commit 0ec09ac2cebe9769491a470c33edff0f873ff79d)

[rebase44(groeck): Picked up final upstream patch to avoid conflicts;
	subject and description now match upstream patch.
	Original subject in chromeos-3.18 was "CHROMIUM: devfreq:
	Fix type of frequency table".]
Signed-off-by: Guenter Roeck <groeck@chromium.org>