UPSTREAM: Bluetooth: Perform HCI update for power on synchronously
The request to update HCI during power on is always coming either from
hdev->req_workqueue or through an ioctl, so it's safe to use
hci_req_sync for it. This way we also eliminate potential races with
incoming mgmt commands or other actions while powering on.

Part of this refactoring is the splitting of mgmt_powered() into
mgmt_power_on() and __mgmt_power_off() functions. The main reason is
the different requirements as far as hdev locking is concerned, as
highlighted with the __ prefix of the power off API.

Since the power on in the case of clearing the AUTO_OFF flag cannot be
done synchronously in the set_powered mgmt handler, the hci_power_on
work callback is extended to cover this (which also simplifies the
set_powered helper a lot).

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 2ff13894cfb877cb3d02d96a8402202f0a6f3efd)

[briannorris: 4.4 rebase. Cherry-pick from bluetooth-next, up to
    6d5d2ee63cee ("Bluetooth: add LED trigger for indicating HCI is powered up")
]

Change-Id: I5e8872bc888aa7eacac5370005d578ae0f4c17f3
Signed-off-by: Brian Norris <briannorris@chromium.org>