UPSTREAM: ARM: dts: rockchip: add startup delay to rk3288-veyron panel-regulators
The panels need a bit of time to actually turn on. If this isn't
observed, this results in problems when trying talk to the panels
and thus produces detection errors. 100ms seem to be a safe value
for the time being.

BUG=chrome-os-partner:50756
TEST=Build and boot on rk3288

Change-Id: Ie873dd16b919ab94454c9254cf9dbab8c202c091
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Tested-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org mmind/linux-rockchip v4.7-armsoc/dts32
 commit 1f45e8c6d0161f044d679f242fe7514e2625af4a)