UPSTREAM: Bluetooth: Use continuous scanning when creating LE connections
All LE connections are now triggered through a preceding passive scan
and waiting for a connectable advertising report. This means we've got
the best possible guarantee that the device is within range and should
be able to request the controller to perform continuous scanning. This
way we minimize the risk that we miss out on any advertising packets.

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Cc: stable@vger.kernel.org # 4.3+
(cherry picked from commit 2f99536a5b34d5b0f54723067d68f6cef3f0fdc6)

[briannorris: 4.4 rebase. Cherry-pick from bluetooth-next, up to
    6d5d2ee63cee ("Bluetooth: add LED trigger for indicating HCI is powered up")
]

Change-Id: I7ee2ec68267cc4adc61b946c32c44ca905f764dd
Signed-off-by: Brian Norris <briannorris@chromium.org>