UPSTREAM: Bluetooth: Don't treat connection timeout as a failure
When we're doing background scanning and connection attempts it's
possible we timeout trying to connect and go back to scanning again.
The timeout triggers a HCI_LE_Create_Connection_Cancel which will
trigger a Connection Complete with "Unknown Connection Identifier"
error status. Since we go back to scanning this isn't really a failure
and shouldn't be presented as such to user space through mgmt.

The exception to this is if the connection attempt was due to an
explicit request on an L2CAP socket (indicated by
params->explicit_connect being true). Since the socket will get an
error it's consistent to also notify the failure on mgmt in this case.

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit acb9f911ea1f828822001d72b21f7cc06e6718c7)

[briannorris: 4.4 rebase. Cherry-pick from bluetooth-next, up to
    6d5d2ee63cee ("Bluetooth: add LED trigger for indicating HCI is powered up")
]

Change-Id: Ic425de5cd6a265a7e50ed5227c6b221fca31407b
Signed-off-by: Brian Norris <briannorris@chromium.org>