CHROMIUM: workaround for Infineon TPM defensive timeout (v2)
The SLB9635 TPM by Infineon goes into a self-defense mode when
power is cut in the middle of a command too many times, and getting
it out of that mode is difficult and takes a long time.  This patch
adds a shutdown-time callback that grabs the mutex wrapping a
request-response transaction, so it's less likely that we'll lose
power in the middle of a command.

Note that the TPM may lose power also across suspend/resume (depending
on hardware).  However, the suspend callback sends a SaveState command,
which can only be sent when other running commands have completed.
I am not sure if a command can start after the suspend callback
has completed (I doubt it) but even if it does, it will be
a sufficiently rare event, and the self-defense can tolerate that.

The "gentle shutdown" message is added to make it easier to detect
if we neglect to carry this patch forward.

BUG=chromium:234796
TEST=extensively tested with the librarian simulator
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I1bb0a6d574c96a33f4344d47e3e7584e3f67fb14
Reviewed-on: https://gerrit.chromium.org/gerrit/58902
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
[benzh: 3.14 rebase. tpm.c has been split into multiple files.]
Signed-off-by: Ben Zhang <benzh@chromium.org>
[bfreed: Commit the change reverted to avoid conflicts in upstream patches.]
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316574
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>

[rebase44(groeck): Change moved from drivers/char/tpm/tpm_tis_common.c
	to drivers/char/tpm/tpm_tis.c]
Signed-off-by: Guenter Roeck <groeck@chromium.org>