UPSTREAM: arm64: perf: add format entry to describe event -> config mapping
It's all very well providing an events directory to userspace that
details our events in terms of "event=0xNN", but if we don't define how
to encode the "event" field in the perf attr.config, then it's a waste
of time.

This patch adds a single format entry to describe that the event field
occupies the bottom 10 bits of our config field on ARMv8 (PMUv3).

BUG=chrome-os-partner:54889
TEST=Build and boot on kevin;

Change-Id: I71f9ebf92cd2f7083c10f20a8707a91d4517cbcb
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
(cherry picked from commit 57d74123954dfe70fad12724d19f743ed14cec90)
Reviewed-on: https://chromium-review.googlesource.com/358005
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>